requirements.txt
Django>=4.2.0
Pillow>=9.5.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
paypalrestsdk>=1.13.1
python-dotenv>=1.0.0
django-allauth>=0.54.0
django-countries>=7.5.1
stripe>=5.4.0
django-widget-tweaks==1.5.0 